+++ src/test/groovy/ApmBasePipelineTest.groovy
@@ -403,7 +403,6 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
           build: {},
           buildStatus: currentBuild.currentResult,
           changeSet: [],
-          log: '',
           stepsErrors: [],
           testsErrors: [],
           testsSummary: []
@@ -536,6 +535,10 @@ class ApmBasePipelineTest extends DeclarativePipelineTest {
       }
     })
     helper.registerAllowedMethod('notifyBuildResult', [], null)
+    helper.registerAllowedMethod('obltGitHubComments', [], {
+      def script = loadScript('vars/obltGitHubComments.groovy')
+      return script.call()
+    })
     helper.registerAllowedMethod('preCommitToJunit', [Map.class], null)
     helper.registerAllowedMethod('publishHTML', [Map.class],  null)
     helper.registerAllowedMethod('randomNumber', [Map.class], { m -> return m.min })