+++ src/test/groovy/NotificationManagerStepTests.groovy
@@ -15,6 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+import co.elastic.mock.RunMock
+import co.elastic.mock.RunWrapperMock
 import org.junit.Before
 import org.junit.Test
 import static org.junit.Assert.assertTrue
@@ -28,8 +30,10 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
   void setUp() throws Exception {
     super.setUp()
     script = loadScript('src/co/elastic/NotificationManager.groovy')
-    f = new File("src/test/resources/console-100-lines.log")
     env.TEST = "test"
+    def runBuilding = new RunMock(building: true)
+    def build = new RunWrapperMock(rawBuild: runBuilding, number: 1, result: 'SUCCESS')
+    binding.setVariable('currentBuild', build)
   }
 
   @Test
@@ -41,7 +45,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       testsSummary: readJSON(file: "tests-summary.json"),
       changeSet: readJSON(file: "changeSet-info.json"),
       statsUrl: "https://ecs.example.com/app/kibana",
-      log: f.getText(),
       testsErrors: readJSON(file: "tests-errors.json"),
       stepsErrors: readJSON(file: "steps-errors.json")
     )
@@ -73,7 +76,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       testsSummary: readJSON(file: "tests-summary.json"),
       changeSet: readJSON(file: "changeSet-info.json"),
       statsUrl: "https://ecs.example.com/app/kibana",
-      log: f.getText(),
       testsErrors: readJSON(file: "tests-errors.json"),
       stepsErrors: readJSON(file: "steps-errors.json")
     )
@@ -92,7 +94,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
         testsSummary: readJSON(file: "tests-summary.json"),
         changeSet: readJSON(file: "changeSet-info.json"),
         statsUrl: "https://ecs.example.com/app/kibana",
-        log: f.getText(),
         testsErrors: readJSON(file: "tests-errors.json"),
         stepsErrors: readJSON(file: "steps-errors.json")
       )
@@ -109,7 +110,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
         testsSummary: readJSON(file: "tests-summary.json"),
         changeSet: readJSON(file: "changeSet-info.json"),
         statsUrl: "https://ecs.example.com/app/kibana",
-        log: f.getText(),
         testsErrors: readJSON(file: "tests-errors.json"),
         stepsErrors: readJSON(file: "steps-errors.json")
       )
@@ -126,7 +126,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
         testsSummary: readJSON(file: "tests-summary.json"),
         changeSet: readJSON(file: "changeSet-info.json"),
         statsUrl: "https://ecs.example.com/app/kibana",
-        log: f.getText(),
         testsErrors: readJSON(file: "tests-errors.json"),
         stepsErrors: readJSON(file: "steps-errors.json")
       )
@@ -140,7 +139,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       buildStatus: "SUCCESS",
       changeSet: readJSON(file: "changeSet-info.json"),
       docsUrl: 'foo',
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -161,7 +159,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       script.notifyPR(
         build: readJSON(file: "build-info.json"),
         changeSet: readJSON(file: "changeSet-info.json"),
-        log: f.getText(),
         statsUrl: "https://ecs.example.com/app/kibana",
         stepsErrors: readJSON(file: "steps-errors.json"),
         testsErrors: readJSON(file: "tests-errors.json"),
@@ -177,7 +174,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       script.notifyPR(
         buildStatus: "FAILURE",
         changeSet: readJSON(file: "changeSet-info.json"),
-        log: f.getText(),
         statsUrl: "https://ecs.example.com/app/kibana",
         stepsErrors: readJSON(file: "steps-errors.json"),
         testsErrors: readJSON(file: "tests-errors.json"),
@@ -191,7 +187,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "ABORTED",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -209,7 +204,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info_aborted.json"),
       buildStatus: "ABORTED",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -228,7 +222,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "FAILURE",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -247,7 +240,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       buildStatus: "UNSTABLE",
       changeSet: readJSON(file: "changeSet-info.json"),
       docsUrl: 'foo',
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -267,7 +259,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "UNSTABLE",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -284,7 +275,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "UNSTABLE",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors-with-multiline.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -301,7 +291,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "UNSTABLE",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors-with-multiple.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -319,7 +308,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "UNSTABLE",
       changeSet: [],
-      log: f.getText(),
       stepsErrors: [],
       testsErrors: readJSON(file: "tests-errors-with-long-stacktrace.json"),
       testsSummary: readJSON(file: 'tests-summary.json')
@@ -338,7 +326,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "UNKNOWN",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -355,7 +342,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "foo",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -374,7 +360,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       changeSet: readJSON(file: "changeSet-info.json"),
       disableGHComment: true,
       docsUrl: 'foo',
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors.json"),
       testsErrors: readJSON(file: "tests-errors.json"),
@@ -393,7 +378,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       buildStatus: "SUCCESS",
       changeSet: readJSON(file: "corrupted/changeSet-info.json"),
       docsUrl: 'foo',
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "corrupted/steps-errors.json"),
       testsErrors: readJSON(file: "corrupted/tests-errors.json"),
@@ -414,7 +398,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "FAILURE",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors-with-github-environmental-issue.json"),
       testsErrors: [],
@@ -432,7 +415,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "FAILURE",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors-with-deleteDir-issue.json"),
       testsErrors: [],
@@ -460,7 +442,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
                     build: readJSON(file: "build-info.json"),
                     buildStatus: "FAILURE",
                     changeSet: readJSON(file: "changeSet-info.json"),
-                    log: f.getText(),
                     statsUrl: "https://ecs.example.com/app/kibana",
                     stepsErrors: readJSON(file: "steps-errors-with-github-environmental-issue.json"),
                     testsErrors: [],
@@ -476,7 +457,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       build: readJSON(file: "build-info.json"),
       buildStatus: "FAILURE",
       changeSet: readJSON(file: "changeSet-info.json"),
-      log: f.getText(),
       statsUrl: "https://ecs.example.com/app/kibana",
       stepsErrors: readJSON(file: "steps-errors-with-also-github-environmental-issue.json"),
       testsErrors: [],
@@ -948,7 +928,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       buildStatus: 'SUCCESS',
       changeSet: readJSON(file: 'changeSet-info.json'),
       docsUrl: 'foo',
-      log: f.getText(),
       statsUrl: 'https://ecs.example.com/app/kibana',
       stepsErrors: readJSON(file: 'steps-errors.json'),
       testsErrors: readJSON(file: 'tests-errors.json'),
@@ -958,7 +937,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
     assertTrue(assertMethodCallContainsPattern('libraryResource', 'github-comment-markdown.template'))
     assertTrue(assertMethodCallContainsPattern('writeFile', 'badge/docs-preview'))
     assertTrue(assertMethodCallContainsPattern('writeFile', 'Build Succeeded'))
-    assertTrue(assertMethodCallContainsPattern('writeFile', 'Image of Build Times'))
     assertTrue(assertMethodCallContainsPattern('writeFile', 'file=build.md'))
     assertTrue(assertMethodCallContainsPattern('archiveArtifacts', 'build.md'))
     assertJobStatusSuccess()
@@ -971,7 +949,6 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
       buildStatus: 'SUCCESS',
       changeSet: readJSON(file: 'changeSet-info.json'),
       docsUrl: 'foo',
-      log: f.getText(),
       statsUrl: 'https://ecs.example.com/app/kibana',
       stepsErrors: readJSON(file: 'steps-errors.json'),
       testsErrors: readJSON(file: 'tests-errors.json'),
@@ -1019,5 +996,13 @@ class NotificationManagerStepTests extends ApmBasePipelineTest {
     printCallStack()
     assertTrue(assertMethodCallContainsPattern('githubPrComment', 'message=awesome'))
     assertJobStatusSuccess()
+  }
+
+  @Test
+  void test_notifyGitHubCommandsInPR_with_empty() throws Exception {
+    helper.registerAllowedMethod('isPR', { return true })
+    script.notifyGitHubCommandsInPR()
+    printCallStack()
+    assertFalse(assertMethodCallContainsPattern('githubPrComment', "GitHub comments"))
   }
 }